+++ include/rain/random.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "functional.hpp"
+
 #include <chrono>
 #include <random>
 
@@ -11,4 +13,62 @@ namespace Rain::Random {
 		std::chrono::duration_cast<std::chrono::nanoseconds>(
 			std::chrono::high_resolution_clock::now().time_since_epoch())
 			.count());
+
+	// 64-bit hash from <https://codeforces.com/blog/entry/62393>.
+	template <typename Type>
+	struct SplitMixHash {
+		std::size_t operator()(Type const &value) const {
+			static const std::size_t FIXED_RANDOM(
+				std::chrono::duration_cast<std::chrono::nanoseconds>(
+					std::chrono::high_resolution_clock::now().time_since_epoch())
+					.count());
+			std::size_t hash{std::hash<Type>{}(value)};
+			hash += FIXED_RANDOM + 0x9e3779b97f4a7c15;
+			hash = (hash ^ (hash >> 30)) * 0xbf58476d1ce4e5b9;
+			hash = (hash ^ (hash >> 27)) * 0x94d049bb133111eb;
+			return hash ^ (hash >> 31);
+		}
+	};
+
+	// Simple function to combine two 32 or 64-bit hashes, based on
+	// <https://stackoverflow.com/questions/5889238/why-is-xor-the-default-way-to-combine-hashes>
+	// from Boost.
+	template <typename Type, typename HashFunctor = SplitMixHash<Type>>
+	inline void combineHash(
+		std::size_t &seed,
+		Type const &value,
+		HashFunctor hasher = SplitMixHash<Type>{}) {
+		if constexpr (sizeof(size_t) >= 8) {
+			seed ^= hasher(value) + 0x517cc1b727220a95 + (seed << 6) + (seed >> 2);
+		} else {
+			seed ^= hasher(value) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+		}
+	}
+
+	// Custom hash functor for const-iterable containers. Cannot be placed in std
+	// namespace as it is UB for system types (not UB for user-defined types).
+	template <
+		typename Container,
+		typename std::enable_if<Functional::isConstIterable<Container>::value>::type
+			* = nullptr>
+	struct ContainerHash {
+		std::size_t operator()(Container const &value) const {
+			std::size_t result{0};
+			for (auto const &i : value) {
+				combineHash(result, i);
+			}
+			return result;
+		}
+	};
+
+	// Custom hash functor for std::pair (TODO: implement tuple functor).
+	template <typename TypeFirst, typename TypeSecond>
+	struct PairHash {
+		std::size_t operator()(
+			std::pair<TypeFirst, TypeSecond> const &value) const {
+			std::size_t result{std::hash<TypeFirst>{}(value.first)};
+			combineHash(result, value.second);
+			return result;
+		}
+	};
 }